% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_initialpop.R
\name{generating_initialpop}
\alias{generating_initialpop}
\title{Generating initial population}
\usage{
generating_initialpop(
  nstudies,
  action_mode,
  par_names,
  lower_bounds,
  upper_bounds,
  model_version = c("TMM", "PIM")
)
}
\arguments{
\item{nstudies}{Numeric, number of starting values in the LHS.}

\item{action_mode}{Character, action mode (executed or imagined).}

\item{par_names}{Character, vector of parameter names.}

\item{lower_bounds}{Numeric, vector of lower bounds for parameters.}

\item{upper_bounds}{Numeric, vector of upper bounds for parameters.}

\item{model_version}{Character, threshold modulation model ("TMM") or parallel inhibition model ("PIM").}
}
\value{
...
}
\description{
Generating plausible initial parameter values for DEoptim based on empirical constraints.
}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
