% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss}
\alias{loss}
\title{Loss functions}
\usage{
loss(
  par,
  data,
  nsims = NULL,
  nsamples = 3000,
  model_version = c("TMM3", "TMM4", "PIM"),
  exec_threshold = 1,
  imag_threshold = 0.5,
  uncertainty = c("par_level", "func_level", "diffusion"),
  time_step = 0.001,
  error_function = c("g2", "rmse", "sse", "wsse", "ks")
)
}
\arguments{
\item{par}{Numeric, vector of parameter values.}

\item{data}{Dataframe, data to be used for computing the error.}

\item{nsims}{number of simulations (observations/trials).}

\item{nsamples}{number of samples (time steps) within a trial.}

\item{model_version}{Version of the model ("TMM" or "PIM").}

\item{exec_threshold}{motor execution threshold.}

\item{imag_threshold}{motor imagery threshold.}

\item{uncertainty}{Numeric, indicates how noise is introduced in the system.}

\item{time_step}{Numeric, time step used to numerical approximation.}

\item{error_function}{Character, loss function to be used when fitting the model.}
}
\value{
The loss/error for a given set of observations.
}
\description{
Provides possible loss functions for the two versions of the model.
}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
