% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momimi-package.R
\docType{package}
\name{momimi-package}
\alias{momimi}
\alias{momimi-package}
\title{momimi: Models of Motor Inhibition during Motor Imagery}
\description{
Provides source code for the "threshold modulation model" (TMM) and the "parallel inhibition model" (PIM) of motor inhibition during motor imagery as well as fitting and plotting utilities.
}
\author{
\strong{Maintainer}: Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com} (\href{https://orcid.org/0000-0002-7419-9855}{ORCID})

Other contributors:
\itemize{
  \item Thibault Gajdos \email{thibault.gajdos@gmail.com} [contributor]
}

}
\keyword{internal}
