% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset_offset.R
\name{onset_offset}
\alias{onset_offset}
\title{Finding onset and offset}
\usage{
onset_offset(peak_time = 0.5, curvature = 0.4, thresh = 0.5)
}
\arguments{
\item{peak_time}{Numeric, peak time of the activation function.}

\item{curvature}{Numeric, curvature of the activation function.}

\item{thresh}{Numeric, a threshold defining onset and offset.}
}
\value{
A vector containing the onset (i.e., the lowest value above the threshold) and the offset (i.e., the highest value above the threshold).
}
\description{
Finding onset and offset of the activation function (according to the model and some threshold).
}
\examples{
\dontrun{
onset_offset(peak_time = 0.5, curvature = 0.4, thresh = 0.5)
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
