% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_props.R
\name{quantiles_props}
\alias{quantiles_props}
\title{Find quantile proportions}
\usage{
quantiles_props(x, quants)
}
\arguments{
\item{x}{Numeric, number of starting values in the LHS.}

\item{quants}{Character, action mode (executed or imagined).}
}
\value{
Vector of proportions per quantile.
}
\description{
Computes the proportion of observations within quantiles
}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
