% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\name{activation}
\alias{activation}
\title{Activation function}
\usage{
activation(
  time = 0,
  amplitude = 1.5,
  peak_time = 0,
  curvature = 0.4,
  uncertainty = c("par_specific", "overall", "brownian")
)
}
\arguments{
\item{time}{Numeric, time within a trial (in seconds).}

\item{amplitude}{Numeric, amplitude of the activation function.}

\item{peak_time}{Numeric, peak time of the activation function.}

\item{curvature}{Numeric, curvature of the activation function.}

\item{uncertainty}{Numeric, indicates how noise is introduced in the system.}
}
\value{
A dataframe
}
\description{
Defining the activation/inhibition rescaled lognormal function.
}
\examples{
\dontrun{
# plotting the activation/inhibition function
curve(
    expr = activation,
    from = 0, 5,
    main = "Activation curve",
    xlab = "Time (in seconds)",
    ylab = "Activation (arbitrary units)"
    )
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
