% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitting}
\alias{fitting}
\title{Fitting the model}
\usage{
fitting(
  par,
  data,
  nsims = NULL,
  par_names = NULL,
  lower_bounds,
  upper_bounds,
  nstudies = 200,
  initial_pop_constraints = FALSE,
  error_function,
  model_version,
  method = c("SANN", "GenSA", "pso", "hydroPSO", "DEoptim", "Nelder-Mead", "BFGS",
    "L-BFGS-B", "bobyqa", "nlminb", "all_methods", "optimParallel"),
  maxit = 100
)
}
\arguments{
\item{par}{Numeric, vector of parameter values.}

\item{data}{Dataframe, data to be used for fitting the model.}

\item{nsims}{Numeric, number of studies to be simulated.}

\item{par_names}{Character, vector of parameter names.}

\item{lower_bounds}{Numeric, vector of lower bounds for parameters.}

\item{upper_bounds}{Numeric, vector of upper bounds for parameters.}

\item{nstudies}{Numeric, number of starting values in the LHS.}

\item{initial_pop_constraints}{Boolean, whether to use additional constraints when sampling initial parameter values.}

\item{error_function}{Character, error function to be used when fitting the model.}

\item{model_version}{Version of the model ("TMM" or "PIM").}

\item{method}{Optimisation method.}

\item{maxit}{Maximum number of iterations.}
}
\value{
The optimised parameter values and further convergence informatio
}
\description{
Fitting the two versions of the model.
}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
