% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitting}
\alias{fitting}
\title{Fitting the model}
\usage{
fitting(
  par,
  data,
  nsims = NULL,
  par_names = NULL,
  lower_bounds,
  upper_bounds,
  nstudies = 200,
  initial_pop_constraints = FALSE,
  rt_contraints = c(0.1, 2),
  mt_contraints = c(0.1, 2),
  error_function = c("g2", "rmse", "sse", "wsse", "ks"),
  uncertainty = c("par_level", "func_level", "diffusion"),
  method = c("SANN", "GenSA", "pso", "DEoptim", "Nelder-Mead", "BFGS", "L-BFGS-B",
    "bobyqa", "nlminb", "all_methods", "optimParallel", "grid_search"),
  cluster = NULL,
  grid_resolution = 0.01,
  maxit = 100,
  verbose = TRUE
)
}
\arguments{
\item{par}{Numeric, vector of parameter values.}

\item{data}{Dataframe, data to be used for fitting the model.}

\item{nsims}{Numeric, number of studies to be simulated.}

\item{par_names}{Character, vector of parameter names.}

\item{lower_bounds}{Numeric, vector of lower bounds for parameters.}

\item{upper_bounds}{Numeric, vector of upper bounds for parameters.}

\item{nstudies}{Numeric, number of starting values in the LHS.}

\item{initial_pop_constraints}{Boolean, whether to use additional constraints when sampling initial parameter values.}

\item{rt_contraints}{Numeric, vector of length 2 specifying the min and max RT (in seconds).}

\item{mt_contraints}{Numeric, vector of length 2 specifying the min and max MT (in seconds).}

\item{error_function}{Character, error function to be used when fitting the model.}

\item{uncertainty}{Numeric, indicates how noise is introduced in the system.}

\item{method}{Character, optimisation method (DEoptim seems to work best).}

\item{cluster}{Character, existing parallel cluster object. If provided, overrides + specified parallelType.}

\item{grid_resolution}{Numeric, resolution of the grid when method = "grid_search".}

\item{maxit}{Numeric, maximum number of iterations.}

\item{verbose}{Boolean, whether to print progress during fitting.}
}
\value{
The optimised parameter values and further convergence information.
}
\description{
Fitting the "threshold modulation model" (TMM) of motor inhibition during motor imagery.
}
\examples{
\dontrun{
plausible "true" parameter values in the TMM (with 4 free parameters)
true_pars <- c(1.1, 0.5, 0.3, 1.25)

# simulating data using these parameter values
simulated_data <- simulating(
    nsims = 200,
    nsamples = 2000,
    true_pars = true_pars,
    action_mode = "imagined"
    )

# fitting the model
# to keep a parameter constant, assign the same value to both upper and lower...
results <- fitting(
    data = simulated_data,
    nsims = 200,
    error_function = "g2",
    method = "DEoptim",
    lower_bounds = c(0, 0.25, 0.1, 1),
    upper_bounds = c(2, 1.25, 0.6, 2),
    initial_pop_constraints = TRUE,
    maxit = 100
    )

# fitting summary
summary(results)
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
