% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_initialpop.R
\name{generating_initialpop}
\alias{generating_initialpop}
\title{Generating initial population}
\usage{
generating_initialpop(
  nstudies,
  action_mode,
  par_names,
  lower_bounds,
  upper_bounds,
  model_version = c("TMM3", "TMM4", "PIM"),
  rt_contraints = c(0.1, 2),
  mt_contraints = c(0.1, 2),
  verbose = TRUE
)
}
\arguments{
\item{nstudies}{Numeric, number of starting values in the LHS.}

\item{action_mode}{Character, action mode (executed or imagined).}

\item{par_names}{Character, vector of parameter names.}

\item{lower_bounds}{Numeric, vector of lower bounds for parameters.}

\item{upper_bounds}{Numeric, vector of upper bounds for parameters.}

\item{model_version}{Character, threshold modulation model ("TMM3" or "TMM4") or parallel inhibition model ("PIM").}

\item{rt_contraints}{Numeric, vector of length 2 specifying the min and max RT (in seconds).}

\item{mt_contraints}{Numeric, vector of length 2 specifying the min and max MT (in seconds).}

\item{verbose}{Boolean, whether to print progress during fitting.}
}
\value{
Returns a dataframe containing plausible (according to custom constraints) initial parameter values.
}
\description{
Generating plausible initial parameter values for DEoptim based on empirical constraints.
}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
