% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Models of motor inhibition during motor imagery}
\usage{
model(
  nsims = 100,
  nsamples = 3000,
  exec_threshold = 1,
  imag_threshold = 0.5,
  peak_time = 0.5,
  curvature = 0.6,
  bw_noise = NULL,
  uncertainty = c("par_level", "func_level", "diffusion"),
  full_output = FALSE,
  time_step = 0.001
)
}
\arguments{
\item{nsims}{Numeric, number of simulations (observations/trials).}

\item{nsamples}{Numeric, number of samples (time steps) within a trial.}

\item{exec_threshold}{Numeric, motor execution threshold.}

\item{imag_threshold}{Numeric, motor imagery threshold.}

\item{peak_time}{Numeric, peak time of the activation function.}

\item{curvature}{Numeric, curvature of the activation function.}

\item{bw_noise}{Numeric, amount of between-trial noise.}

\item{uncertainty}{Numeric, indicates how noise is introduced in the system.}

\item{full_output}{Boolean, indicating whether activation curves should be returned.}

\item{time_step}{Numeric, time step used to numerical approximation.}
}
\value{
A dataframe containing observation (i.e., RTs and MTs) and/or values of the underlying functions at each time step.
}
\description{
Simulating data from the "threshold modulation model" (TMM) of motor imagery.
}
\examples{
\dontrun{
# simulating 500 observations (pairs of RT and MT)
simulated_data <- model(
    nsims = 500,
    exec_threshold = 1,
    peak_time = log(0.5),
    curvature = 0.4,
    bw_noise = 0.08,
    full_output = FALSE
    )

# displaying the first six observations
head(simulated_data)

# plotting the distributions of RTs/MTs
plot(simulated_data)
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
