% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Models of motor inhibition during motor imagery}
\usage{
model(
  nsims = 100,
  nsamples = 3000,
  exec_threshold = 1,
  imag_threshold = 0.5,
  amplitude_activ = 1.5,
  peak_time_activ = 0,
  curvature_activ = 0.4,
  amplitude_inhib = 1.5,
  peak_time_inhib = 0,
  curvature_inhib = 0.6,
  bw_noise = NULL,
  model_version = c("TMM3", "TMM4", "PIM"),
  uncertainty = c("par_specific", "brownian", "overall"),
  full_output = FALSE,
  time_step = 0.001
)
}
\arguments{
\item{nsims}{Numeric, number of simulations (observations/trials).}

\item{nsamples}{Numeric, number of samples (time steps) within a trial.}

\item{exec_threshold}{Numeric, motor execution threshold.}

\item{imag_threshold}{Numeric, motor imagery threshold.}

\item{amplitude_activ}{Numeric, amplitude of the activation function.}

\item{peak_time_activ}{Numeric, peak time of the activation function.}

\item{curvature_activ}{Numeric, curvature of the activation function.}

\item{amplitude_inhib}{Numeric, amplitude of the inhibition function.}

\item{peak_time_inhib}{Numeric, peak time of the inhibition function.}

\item{curvature_inhib}{Numeric, curvature of the inhibition function.}

\item{bw_noise}{Numeric, amount of between-trial noise.}

\item{model_version}{Character, threshold modulation model ("TMM3" or "TMM4") or parallel inhibition model ("PIM").}

\item{uncertainty}{Numeric, indicates how noise is introduced in the system.}

\item{full_output}{Boolean, indicating whether activ/inhib curves should be returned.}

\item{time_step}{Numeric, time step used to numerical approximation.}
}
\value{
A dataframe containing observation (i.e., RTs and MTs) and/or values of the underlying functions at each time step.
}
\description{
Sampling from the two versions of the model.
}
\examples{
\dontrun{
# simulating 100 observations
simulated_data <- model(
    nsims = 100, nsamples = 2000,
    exec_threshold = 1, imag_threshold = 0.5,
    amplitude_activ = 0.8, peak_time_activ = log(0.5), curvature_activ = 0.4,
    amplitude_inhib = 1.75, peak_time_inhib = log(0.5), curvature_inhib = 0.5,
    model_version = "PIM",
    full_output = FALSE
    )

# displaying the first six observations
head(simulated_data)
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
