% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{simulating}
\alias{simulating}
\title{Simulating data}
\usage{
simulating(
  nsims = 100,
  nsamples = 2000,
  true_pars = NULL,
  action_mode = c("executed", "imagined"),
  model_version = c("TMM", "PIM")
)
}
\arguments{
\item{nsims}{Numeric, number of studies to be simulated.}

\item{nsamples}{Numeric, number of samples (time steps) within a trial.}

\item{true_pars}{Numeric, vector of "true" parameter values.}

\item{action_mode}{Character, whether to simulate executed or imagined trials.}

\item{model_version}{Character, version of the model ("TMM" or "PIM").}
}
\value{
A dataframe containing observation (i.e., RTs and MTs).
}
\description{
Simulating data from the "threshold modulation model" (TMM) or the "parallel inhibition model" (PIM) of motor inhibition during motor imagery.
}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
