% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{simulating}
\alias{simulating}
\title{Simulating data}
\usage{
simulating(
  nsims = 100,
  nsamples = 3000,
  true_pars = NULL,
  action_mode = c("executed", "imagined"),
  uncertainty = c("par_level", "func_level", "diffusion"),
  bw_noise = NULL
)
}
\arguments{
\item{nsims}{Numeric, number of studies to be simulated.}

\item{nsamples}{Numeric, number of samples (time steps) within a trial.}

\item{true_pars}{Numeric, vector of "true" parameter values.}

\item{action_mode}{Character, whether to simulate executed or imagined trials.}

\item{uncertainty}{Numeric, indicates how noise is introduced in the system.}

\item{bw_noise}{Numeric, amount of between-trial noise.}
}
\value{
A dataframe containing observation (i.e., RTs and MTs).
}
\description{
Simulating data from the "threshold modulation model" (TMM).
}
\examples{
\dontrun{
# plausible "true" parameter values in the TMM
# parameters are the motor execution threshold, the peak time,
# curvature, and between-trial variability
true_pars <- c(1.2, 0.5, 0.5, 0.06)

# simulating data using these parameter values
simulated_data <- simulating(
    nsims = 200,
    true_pars = true_pars,
    action_mode = "imagined"
    )

# plotting the simulated data
plot(simulated_data)
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
